move code from Upload middleware to FileService
